+++ src/angular-integration.ts
@@ -1,6 +1,9 @@
 import {Observable} from "rxjs/Observable";
+import {Operator} from "rxjs/Operator";
 import {takeUntil} from "rxjs/operators";
 import {Subject} from "rxjs/Subject";
+import {Subscriber} from "rxjs/Subscriber";
+import {operatorFactory} from "./utils";
 
 export type OnDestroyLike = {
     ngOnDestroy(): void;
@@ -19,6 +22,19 @@ export function componentDestroyed(component: OnDestroyLike): Observable<true> {
     return stop$.asObservable();
 }
 
-export function untilComponentDestroyed<T>(component: OnDestroyLike): (source: Observable<T>) => Observable<T> {
-    return (source: Observable<T>) => source.pipe(takeUntil(componentDestroyed(component)));
+export function toNgComponent<T>(component: OnDestroyLike): Operator<T, T> {
+    return operatorFactory(
+        (subscriber: Subscriber<T>, source: Observable<T>) => {
+            const subscription = source
+                .pipe(
+                    takeUntil(componentDestroyed(component))
+                )
+                .subscribe(subscriber);
+
+            return () => {
+                subscription.unsubscribe();
+                subscriber.complete();
+            };
+        }
+    );
 }