+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from "./Store";
-export * from "./UnboundedObservable";
+export * from "./StoreObserver";
 export {Mutators} from "./mutators";
+export * from "./selectMany";
 export * from "./development";
-